// frontend/types/strapi.ts
export interface StrapiMediaFormat {
  name: string;
  hash: string;
  ext: string;
  mime: string;
  path?: string;
  width: number;
  height: number;
  size: number;
  sizeInBytes: number;
  url: string;
}

export interface StrapiMedia {
  id: number;
  documentId: string;
  name: string;
  alternativeText?: string;
  caption?: string;
  width?: number;
  height?: number;
  formats?: {
    thumbnail?: StrapiMediaFormat;
    mobile?: StrapiMediaFormat;
    tablet?: StrapiMediaFormat;
    desktop?: StrapiMediaFormat;
    wide?: StrapiMediaFormat;
    xlarge?: StrapiMediaFormat;
    webp?: StrapiMediaFormat;
    thumbnail_webp?: StrapiMediaFormat;
    mobile_webp?: StrapiMediaFormat;
    tablet_webp?: StrapiMediaFormat;
    desktop_webp?: StrapiMediaFormat;
    wide_webp?: StrapiMediaFormat;
    xlarge_webp?: StrapiMediaFormat;
  };
  hash: string;
  ext: string;
  mime: string;
  size: number;
  url: string;
  previewUrl?: string;
  provider: string;
  provider_metadata?: Record<string, unknown>;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
}